singleton system dev

+++ SpringBootDev/src/main/java/org/springbootdev/core/tool/node/ForestNodeManager.java
@@ -19,7 +19,7 @@ public class ForestNodeManager<T extends INode> {
 	/**
 	 * 森林的父节点ID
 	 */
-	private List<Integer> parentIds = new ArrayList<>();
+	private List<Long> parentIds = new ArrayList<>();
 
 	public ForestNodeManager(List<T> items) {
 		list = items;
@@ -31,7 +31,7 @@ public class ForestNodeManager<T extends INode> {
 	 * @param id 节点ID
 	 * @return 对应的节点对象
 	 */
-	public INode getTreeNodeAT(int id) {
+	public INode getTreeNodeAT(Long id) {
 		for (INode forestNode : list) {
 			if (forestNode.getId() == id) {
 				return forestNode;
@@ -45,7 +45,7 @@ public class ForestNodeManager<T extends INode> {
 	 *
 	 * @param parentId 父节点ID
 	 */
-	public void addParentId(Integer parentId) {
+	public void addParentId(Long parentId) {
 		parentIds.add(parentId);
 	}
