+++ SpringBootDev/src/main/java/org/springbootdev/modules/system/entity/Tenant.java
@@ -13,7 +13,7 @@ import org.springbootdev.core.mp.base.BaseEntity;
 /**
  * 实体类
  *
- * @author merryChen
+ * @author zhaobohao
  */
 @Data
 @TableName("mk_tenant")
@@ -28,7 +28,7 @@ public class Tenant extends BaseEntity {
 	 */
 	@TableId(value = "id", type = IdType.AUTO)
 	@ApiModelProperty(value = "主键id")
-	private Integer id;
+	private Long id;
 
 	/**
 	 * 租户ID
